requirements.txt
beautifulsoup4==4.14.2
bs4==0.0.2
lxml==6.0.2
numpy==2.3.5
pandas==2.3.3
PyYAML==6.0.3
requests==2.32.5